yolo-in-go-with-python/yolo-api/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.11.1
bleach==6.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.2.2
cryptography==41.0.3
cssselect2==0.7.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docarray==0.32.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
hnswlib==0.8.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.47
llvmlite==0.42.0
lxml==4.9.2
markdown-it-py==2.1.0
markdown2==2.4.8
markdownmaker==0.4.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
md2pdf==0.5
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
openai==1.17.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.4.0
pinecone-client==3.2.2
platformdirs==4.2.0
pluggy==1.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.8
pydantic_core==2.18.1
pydyf==0.5.0
pyfiglet==0.8.post1
Pygments==2.14.0
pypdf==4.2.0
pyphen==0.13.2
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-gitlab==3.12.0
python-json-logger==2.0.7
pytube==15.0.0
PyYAML @ file:///private/tmp/pyyaml-20231019-6134-1qsnlp0/PyYAML-6.0.1
pyzmq==26.0.0
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rpds-py==0.18.0
ruff==0.3.7
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
torch==2.2.2
tornado==6.4
tqdm==4.64.1
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
weasyprint==58.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
yarl==1.9.4
zopfli==0.2.2
